reward is float instead of int

+++ Plugins/HolodeckCore/Source/HolodeckCore/Public/HolodeckAgent.h
@@ -28,7 +28,7 @@ public:
 	FString AgentName;
 
 	UPROPERTY(BlueprintReadWrite)
-	int Reward;
+	float Reward;
 
 	UPROPERTY(BlueprintReadWrite)
 	bool Terminal;
